% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrayhelpers-package.R
\docType{package}
\name{arrayhelpers-package}
\alias{arrayhelpers}
\alias{arrayhelpers-package}
\title{arrayhelpers: Convenience Functions for Arrays}
\description{
Some convenient functions to work with arrays.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://arrayhelpers.r-forge.r-project.org/}
}

}
\author{
\strong{Maintainer}: Claudia Beleites \email{Claudia.Beleites@chemometrix.eu}

Authors:
\itemize{
  \item Claudia Beleites \email{Claudia.Beleites@chemometrix.eu}
}

}
\keyword{internal}
