\name{countrows}
\alias{countRows}
\title{Count equal rows...}
\usage{countRows(x)
}
\description{Count equal rows}
\details{matrices are converted to data.frame.}
\value{data frame with unique rows, their counts and indices into the original data.frame}
\note{this function is subject to changes in the future.}
\author{Claudia Beleites}
\arguments{\item{x}{the matrix or data.frame}
}

