\name{ndim}
\alias{ndim}
\title{number of dimensions}
\usage{
  ndim(a)
}
\arguments{
  \item{a}{vector, matrix, or array}

  \item{...}{indexing instructions. The names of the
  arguments specify the dimension (i = 1st, j = 2nd, ...).
  The indexing expressions are the same as for
  \code{\link[base]{[}}}
}
\value{
  integer: length of dim attribute
}
\description{
  number of dimensions
}
\author{
  Claudia Beleites
}

