\name{numericindex}
\alias{numericindex}
\title{Convert character or logical indices to numeric}
\usage{
  numericindex(x, i, n = names(x))
}
\arguments{
  \item{x}{the object that is to be indexed}

  \item{n}{names of the object}

  \item{i}{the indices to be converted}
}
\value{
  numeric indices
}
\description{
  Convert character or logical indices to numeric
}
\author{
  Claudia Beleites
}

