% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowsum.R
\docType{methods}
\name{rowsum,array-method}
\alias{rowsum,array-method}
\title{rowsum for arrays}
\usage{
\S4method{rowsum}{array}(x, group, reorder = TRUE, na.rm = FALSE, ...)
}
\arguments{
\item{x}{array to be \code{rowsum}med}

\item{group}{grouping variable (integer or factor) indicating groups of samples. \code{}}

\item{reorder}{should the groups be ordered? see \code{\link[base]{rowsum}}}

\item{na.rm}{shoud \code{NA}s be removed?}

\item{...}{ignored}
}
\value{
like \code{\link[base]{rowsum}}, but further dimensions of the array are preserved.
}
\description{
This function extends the base function \code{\link[base]{rowsum}}.
}
\author{
Claudia Beleites
}
\seealso{
\code{\link[base]{rowsum}}
}
\keyword{algebra}
\keyword{arith}
\keyword{array}

