% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrayhelpers.R, R/arrayhelpers-package.R
\docType{package}
\name{arrayhelpers-package}
\alias{arrayhelpers-package}
\title{Package arrayhelpers}
\description{
Little helper functions to work with arrays

Some convenient functions to work with arrays.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://arrayhelpers.r-forge.r-project.org/}
}

}
\author{
C. Beleites

Maintainer: Claudia Beleites <claudia.beleites@chemometrix.eu>

\strong{Maintainer}: Claudia Beleites \email{Claudia.Beleites@chemometrix.eu}

Authors:
\itemize{
  \item Claudia Beleites \email{Claudia.Beleites@chemometrix.eu}
}

}
\keyword{internal}
\keyword{package}
