% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrayhelpers.R
\docType{package}
\name{arrayhelpers-package}
\alias{arrayhelpers-package}
\title{Package arrayhelpers}
\description{
Little helper functions to work with arrays
}
\author{
C. Beleites

Maintainer: Claudia Beleites <claudia.beleites@chemometrix.eu>
}
\keyword{package}
