\name{arrayhelpers.unittest}
\alias{arrayhelpers.unittest}
\title{Run the unit tests...}
\description{Run the unit tests}
\details{Run the unit tests attached to the functions via \link[svUnit]{svUnit}}
\value{invisibly \code{TRUE} if the tests pass, \code{NA} if \link[svUnit]{svUnit} is not
available. Stops if errors are encountered.}
\author{Claudia Beleites}
\seealso{\link[svUnit]{svUnit}}

