\name{countRows}
\alias{countRows}
\title{Count equal rows}
\usage{
  countRows(x)
}
\arguments{
  \item{x}{the matrix or data.frame}
}
\value{
  data frame with unique rows, their counts and indices
  into the original data.frame
}
\description{
  matrices are converted to data.frame.
}
\note{
  this function is subject to changes in the future.
}
\author{
  Claudia Beleites
}

