\name{dropdimnames}
\alias{dropdimnames}
\title{Drop dimnames if all elements are NULL...}
\usage{dropdimnames(x)
lon(l)
}
\description{Drop dimnames if all elements are \code{NULL}}
\value{\code{dropdimnames}: object without empty dimnames

\code{lon}: \code{lon}: \code{NULL} if all elements of \code{dn} are \code{NULL}, otherwise \code{dn}


}
\author{Claudia Beleites}
\arguments{\item{x}{object}
\item{l}{list}
}
\alias{lon}

