\name{ensuredim}
\alias{ensuredim}
\title{Ensure dim attribute...}
\usage{ensuredim(v)
}
\description{Ensure dim attribute}
\details{Turns vectors into 1d-arrays, and leaves arrays unchanged.}
\value{vectors are turned in "arrays" of one dimension}
\author{Claudia Beleites}
\arguments{\item{v}{vector (or array)}
}

