% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndim.R
\name{ndim}
\alias{ndim}
\title{number of dimensions}
\usage{
ndim(a)
}
\arguments{
\item{a}{vector, matrix, or array}
}
\value{
integer: length of dim attribute
}
\description{
number of dimensions
}
\author{
Claudia Beleites
}
