% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericindices.R
\name{numericindex}
\alias{numericindex}
\title{Convert character or logical indices to numeric}
\usage{
numericindex(x, i, n = names(x))
}
\arguments{
\item{x}{the object that is to be indexed}

\item{i}{the indices to be converted}

\item{n}{names of the object}
}
\value{
numeric indices
}
\description{
Convert character or logical indices to numeric
}
\author{
Claudia Beleites
}

