\name{numericindex}
\alias{numericindex}
\title{Convert character or logical indices to numeric...}
\usage{numericindex(x, i, n=names(x))
}
\description{Convert character or logical indices to numeric}
\value{numeric indices}
\author{Claudia Beleites}
\arguments{\item{x}{the object that is to be indexed}
\item{n}{names of the object}
\item{i}{the indices to be converted}
}

